remove model and move helper functions out of licensing
