+++ awx/main/utils/candlepin/lifecycle.py
@@ -195,83 +195,3 @@ def get_candlepin_ca():
 def get_proxy_url():
     """Get proxy URL from Django settings."""
     return settings.AWX_ANALYTICS_CANDLEPIN_PROXY_URL
-
-
-def get_certificate_info():
-    """
-    Get detailed certificate information for monitoring and status.
-
-    Returns a dict with certificate status, expiry info, and configuration details.
-    Adapted from PR 16240's certificate_manager.get_certificate_info().
-    """
-    from awx.main.utils.licensing import _fetch_candlepin_cert_from_db
-
-    info = {
-        'status': 'missing',
-        'certificate_auth_enabled': True,
-        'candlepin_url': get_candlepin_url(),
-    }
-
-    cert_pem, key_pem, consumer_uuid = _fetch_candlepin_cert_from_db()
-
-    if consumer_uuid:
-        info['consumer_uuid'] = consumer_uuid
-
-    if not cert_pem or not key_pem:
-        info['message'] = 'No certificate found in database'
-        info['needs_renewal'] = True
-        return info
-
-    try:
-        cert_info = parse_cert(cert_pem)
-        days_until_expiry = cert_info['days_remaining']
-        renewal_threshold = get_renewal_days()
-
-        info['not_valid_after'] = cert_info['not_after']
-        info['days_until_expiry'] = days_until_expiry
-        info['serial_number'] = cert_info['serial']
-        info['cn'] = cert_info['cn']
-
-        if days_until_expiry < 0:
-            info['status'] = 'expired'
-            info['needs_renewal'] = True
-            info['message'] = f'Certificate expired {abs(days_until_expiry)} days ago'
-        elif days_until_expiry <= renewal_threshold:
-            info['status'] = 'expiring_soon'
-            info['needs_renewal'] = True
-            info['message'] = f'Certificate expires in {days_until_expiry} days (threshold: {renewal_threshold} days)'
-        else:
-            info['status'] = 'valid'
-            info['needs_renewal'] = False
-            info['message'] = f'Certificate is valid for {days_until_expiry} more days'
-    except ValueError as e:
-        info['status'] = 'error'
-        info['needs_renewal'] = True
-        info['message'] = f'Certificate parsing failed: {e}'
-
-    return info
-
-
-def check_certificate_health():
-    """
-    Health check for analytics certificate status.
-
-    Returns simplified health status suitable for monitoring systems.
-    Adapted from PR 16240's check_certificate_health().
-    """
-    cert_info = get_certificate_info()
-
-    status_mapping = {
-        'valid': 'healthy',
-        'expiring_soon': 'warning',
-        'expired': 'critical',
-        'missing': 'critical',
-        'error': 'critical',
-    }
-
-    return {
-        'status': status_mapping.get(cert_info.get('status', 'unknown'), 'warning'),
-        'message': cert_info.get('message', f"Certificate status: {cert_info.get('status', 'unknown')}"),
-        'needs_renewal': cert_info.get('needs_renewal', True),
-        'days_until_expiry': cert_info.get('days_until_expiry'),
-    }
