+++ awx/main/utils/licensing.py
@@ -41,17 +41,6 @@ from awx_plugins.interfaces._temporary_private_licensing_api import detect_serve
 
 from awx.main.constants import SUBSCRIPTION_USAGE_MODEL_UNIQUE_HOSTS
 from awx.main.utils.analytics_proxy import OIDCClient
-from awx.main.utils.candlepin.client import CandlepinClient
-from awx.main.utils.candlepin.lifecycle import (
-    get_candlepin_ca,
-    get_candlepin_url,
-    get_proxy_url,
-    get_renewal_days,
-    is_cert_valid,
-    parse_cert,
-    run_candlepin_lifecycle,
-)
-from awx.main.utils.encryption import decrypt_field
 
 MAX_INSTANCES = 9999999
 
@@ -564,257 +553,3 @@ def get_licenser(*args, **kwargs):
             return Licenser(*args, **kwargs)
     except Exception as e:
         raise ValueError(_('Error importing License: %s') % e)
-
-
-# ---------------------------------------------------------------------------
-# Candlepin Integration for mTLS Authentication
-# ---------------------------------------------------------------------------
-
-# Placeholder UUID used before a real consumer is registered.
-# Treat it the same as an absent UUID so we never attempt a Candlepin
-# lifecycle call with a non-functional consumer identity.
-CANDLEPIN_UUID_PLACEHOLDER = '00000000-0000-0000-0000-000000000000'
-
-
-def _fetch_candlepin_cert_from_db():
-    """Read cert PEM, key PEM, and consumer UUID from CandlepinCertificate model.
-
-    Returns (cert_pem, key_pem, consumer_uuid) if valid certificate data exists,
-    or (None, None, None) if the instance only contains placeholder/unregistered data.
-    Best-effort: failures are logged as warnings and never propagate.
-    """
-    from awx.main.models import CandlepinCertificate
-
-    try:
-        instance = CandlepinCertificate.get_instance()
-        # get_instance() always returns a singleton; check if it has real data vs placeholder
-        if not instance.has_valid_data():
-            return None, None, None
-
-        return decrypt_field(instance, 'cert_pem'), decrypt_field(instance, 'key_pem'), instance.consumer_uuid
-    except Exception as e:
-        logger.warning(f'Could not fetch Candlepin lifecycle data from DB: {e}')
-        return None, None, None
-
-
-def _save_candlepin_cert_to_db(cert_pem, key_pem):
-    """Persist a renewed Candlepin identity cert and key to the CandlepinCertificate model.
-
-    Updates the existing instance if present, creates a new one if needed.
-
-    Returns:
-        bool: True if save succeeded, False on any error.
-    """
-    from awx.main.models import CandlepinCertificate
-
-    try:
-        # Get or create the instance
-        instance = CandlepinCertificate.get_or_create_instance()
-
-        # Parse certificate to extract metadata
-        try:
-            cert_info = parse_cert(cert_pem)
-            serial_number = cert_info.get('serial', '')
-            expires_at = datetime.fromisoformat(cert_info['not_after']) if cert_info.get('not_after') else None
-        except Exception as e:
-            logger.warning(f'Could not parse certificate metadata: {e}')
-            serial_number = ''
-            expires_at = None
-
-        # Update the certificate data
-        instance.update_certificate(
-            cert_pem=cert_pem,
-            key_pem=key_pem,
-            serial_number=serial_number,
-            expires_at=expires_at,
-        )
-        logger.info('Renewed Candlepin cert and key saved to database.')
-        return True
-    except Exception as e:
-        logger.error(f'Could not save renewed Candlepin cert to database: {e}')
-        return False
-
-
-def _fetch_registration_credentials_from_db():
-    """Read Candlepin registration credentials from AWX settings.
-
-    Reads REDHAT_USERNAME, REDHAT_PASSWORD (set by AWX when the
-    customer configures their Red Hat subscription), LICENSE.account_number (org
-    key for the Candlepin /consumers endpoint), and INSTALL_UUID (used as the
-    consumer's aap.instance_uuid fact).
-
-    Returns (username, password, org, install_uuid), any of which may be None
-    if the corresponding setting is not configured.
-    """
-    try:
-        username = getattr(settings, 'REDHAT_USERNAME', None)
-        password = getattr(settings, 'REDHAT_PASSWORD', None)
-        install_uuid = getattr(settings, 'INSTALL_UUID', None)
-        license_data = getattr(settings, 'LICENSE', {})
-        org = license_data.get('account_number') if isinstance(license_data, dict) else None
-        return username, password, org, install_uuid
-    except Exception as e:
-        logger.warning(f'Could not fetch Candlepin registration credentials from settings: {e}')
-        return None, None, None, None
-
-
-def _save_candlepin_registration_to_db(cert_pem, key_pem, consumer_uuid):
-    """Persist a new Candlepin consumer registration (cert, key, UUID) to the CandlepinCertificate model.
-
-    Updates the existing instance if present, creates a new one if needed.
-
-    Returns:
-        bool: True if save succeeded, False on any error.
-    """
-    from awx.main.models import CandlepinCertificate
-
-    try:
-        # Get or create the instance
-        instance = CandlepinCertificate.get_or_create_instance()
-
-        # Parse certificate to extract metadata
-        try:
-            cert_info = parse_cert(cert_pem)
-            serial_number = cert_info.get('serial', '')
-            expires_at = datetime.fromisoformat(cert_info['not_after']) if cert_info.get('not_after') else None
-        except Exception as e:
-            logger.warning(f'Could not parse certificate metadata: {e}')
-            serial_number = ''
-            expires_at = None
-
-        # Update the certificate data with new consumer UUID
-        instance.update_certificate(
-            cert_pem=cert_pem,
-            key_pem=key_pem,
-            consumer_uuid=consumer_uuid,
-            serial_number=serial_number,
-            expires_at=expires_at,
-        )
-        logger.info(f'Candlepin consumer registration saved to database (uuid={consumer_uuid}).')
-        return True
-    except Exception as e:
-        logger.error(f'Could not save Candlepin registration to database: {e}')
-        return False
-
-
-def _register_candlepin_consumer():
-    """Register a new Candlepin consumer using credentials from AWX settings.
-
-    Called when no identity cert exists in the DB.
-
-    Reads REDHAT_USERNAME / REDHAT_PASSWORD and the org key from
-    LICENSE.account_number, then calls POST /consumers on Candlepin to obtain an
-    identity certificate.  On success the cert, key, and consumer UUID are
-    persisted to the CandlepinCertificate model.
-
-    Returns (cert_pem, key_pem, consumer_uuid) on success, (None, None, None) on
-    any failure.  Best-effort: logs errors but never propagates.
-    """
-    username, password, org, install_uuid = _fetch_registration_credentials_from_db()
-
-    if not username or not password:
-        logger.warning('Candlepin registration is enabled but REDHAT_USERNAME / REDHAT_PASSWORD are not set; skipping registration.')
-        return None, None, None
-
-    if not org:
-        logger.warning('Candlepin registration is enabled but LICENSE.account_number is not available; skipping registration.')
-        return None, None, None
-
-    candlepin_url = get_candlepin_url()
-    candlepin_ca = get_candlepin_ca()
-    proxy = get_proxy_url()
-    client = CandlepinClient(base_url=candlepin_url, candlepin_ca=candlepin_ca, proxy=proxy)
-
-    try:
-        cert_pem, key_pem, consumer_uuid = client.register_consumer(username, password, org, install_uuid)
-    except Exception as e:
-        logger.error(f'Candlepin consumer registration failed: {e}')
-        return None, None, None
-
-    if not _save_candlepin_registration_to_db(cert_pem, key_pem, consumer_uuid):
-        logger.error('Candlepin consumer registration succeeded but failed to save to database.')
-        return None, None, None
-    return cert_pem, key_pem, consumer_uuid
-
-
-def _run_candlepin_lifecycle(cert_pem, key_pem, consumer_uuid):
-    """Orchestrate Candlepin check-in and proactive cert renewal.
-
-    Returns the (possibly renewed) (cert_pem, key_pem) tuple. If renewal fails, the
-    original cert is returned so the caller can still attempt mTLS (which
-    will then fall back to service-account auth via the existing SSLError
-    handler).
-    """
-    if not consumer_uuid or consumer_uuid == CANDLEPIN_UUID_PLACEHOLDER:
-        logger.warning(
-            'Candlepin lifecycle is enabled but consumer UUID is not set ' '(or still contains the placeholder value); ' 'skipping check-in and renewal.'
-        )
-        return cert_pem, key_pem
-
-    candlepin_url = get_candlepin_url()
-    renewal_days = get_renewal_days()
-    candlepin_ca = get_candlepin_ca()
-    proxy = get_proxy_url()
-
-    try:
-        new_cert_pem, new_key_pem = run_candlepin_lifecycle(
-            cert_pem,
-            key_pem,
-            consumer_uuid,
-            candlepin_url=candlepin_url,
-            renewal_days=renewal_days,
-            candlepin_ca=candlepin_ca,
-            proxy=proxy,
-        )
-        if (new_cert_pem, new_key_pem) != (cert_pem, key_pem):
-            if not _save_candlepin_cert_to_db(new_cert_pem, new_key_pem):
-                logger.warning('Renewed certificate will be used for this request, but failed to persist to database for future use.')
-        return new_cert_pem, new_key_pem
-    except Exception as e:
-        logger.error(f'Candlepin lifecycle (check-in / renewal) failed: {e}; will attempt mTLS with existing cert')
-        return cert_pem, key_pem
-
-
-def get_or_generate_candlepin_certificate(username, password):
-    """
-    Get or generate Candlepin certificate for analytics authentication.
-
-    This function provides certificate-based authentication for analytics uploads.
-    It will:
-    1. Check for existing certificate in the CandlepinCertificate model
-    2. If missing, attempt to register with Candlepin
-    3. If exists, check for renewal needs and refresh if needed
-    4. Return the certificate and key as PEM strings
-
-    Args:
-        username: Red Hat subscription username (REDHAT_USERNAME)
-        password: Red Hat subscription password (REDHAT_PASSWORD)
-
-    Returns:
-        Tuple (cert_pem, key_pem) as strings if certificate is available, (None, None) otherwise.
-
-    Adapted from PR 16240's get_or_generate_client_certificate() but uses Django model
-    instead of file-based storage.
-    """
-    cert_pem, key_pem, consumer_uuid = _fetch_candlepin_cert_from_db()
-
-    # If no certificate exists, attempt registration
-    if not cert_pem or not key_pem:
-        logger.info('No Candlepin certificate found, attempting registration')
-        cert_pem, key_pem, consumer_uuid = _register_candlepin_consumer()
-
-        if not cert_pem or not key_pem:
-            logger.debug('Candlepin certificate registration failed or not configured')
-            return None, None
-
-    # Run lifecycle (check-in and renewal if needed)
-    if consumer_uuid and consumer_uuid != CANDLEPIN_UUID_PLACEHOLDER:
-        cert_pem, key_pem = _run_candlepin_lifecycle(cert_pem, key_pem, consumer_uuid)
-
-    # Validate certificate is still usable
-    if not is_cert_valid(cert_pem):
-        logger.warning('Candlepin certificate is not valid (expired or not yet valid)')
-        return None, None
-
-    # Return raw PEM strings - caller will create temp files if needed
-    return cert_pem, key_pem
